# docs/esp/openapi.yaml
openapi: 3.0.0
info:
  title: ESP - News Pushing Pipeline API
  description: |
    Unified event query API over three data sources:
    - MongoDB (market news from Polygon.io)
    - SQLite (Benzinga earnings calendar)
    - SQLite (Nasdaq economic events calendar)

    ## Event Types
    - `macro_calendar` — Economic calendar events (FOMC, CPI, NFP, etc.)
    - `earnings` — Earnings announcements
    - `breaking_news` — Breaking market news (< 2h old)
    - `daily_news` — General market news

    ## Pagination
    POST endpoints use cursor-based pagination. Pass `next_cursor` from a response
    as `cursor` in the next request to fetch the next page.
  version: 0.1.0
  contact:
    name: QFinZero
servers:
  - url: http://localhost:19702
    description: Local development server
paths:
...
  /esp/news/{news_id}/body:
    get:
      operationId: newsBody
      summary: Get full news article
      tags: [news]
      parameters:
        - name: news_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Full news article
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NewsBodyResponse'
        '404':
          description: Article not found
  /esp/news/search:
    post:
      operationId: searchNews
      summary: Search news with filtering
      description: |
        Search news articles with keyword and publisher filtering.
        Supports cursor-based pagination.
      tags: [news]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewsSearchRequest'
      responses:
        '200':
          description: Paginated search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse'
  /esp/news/stats:
    get:
      operationId: newsStats
      summary: News statistics
      description: |
        Returns news statistics including daily counts, top tickers,
        top publishers, and duplicate detection rates.
      tags: [news]
      parameters:
        - name: days
          in: query
          required: false
          description: Lookback window in days
          schema:
            type: integer
            minimum: 1
            maximum: 90
            default: 7
      responses:
        '200':
          description: News statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NewsStatsResponse'
        '503':
          description: MongoDB unavailable
...
components:
  schemas:
...
    Event:
      type: object
      properties:
        event_id:
          type: string
        event_type:
          type: string
          enum: [macro_calendar, earnings, breaking_news, daily_news]
        title:
          type: string
        time_utc:
          type: string
          format: date-time
        importance:
          type: string
          enum: [low, medium, high]
        status:
          type: string
          enum: [scheduled, occurred, updated]
        tickers:
          type: array
          items:
            type: string
        country:
          type: string
          default: "US"
        snippet:
          type: string
        payload:
          type: object
        source:
          type: string
        source_id:
          type: string

    # ── Request Models ────────────────────────────────────────────
...
    NewsSearchRequest:
      type: object
      properties:
        tickers:
          type: array
          items:
            type: string
        start_utc:
          type: string
          format: date-time
        end_utc:
          type: string
          format: date-time
        keyword:
          type: string
          description: Case-insensitive substring search in title
        publisher:
          type: string
          description: Case-insensitive substring search in publisher name
        limit:
          type: integer
          minimum: 1
          maximum: 500
          default: 50
        cursor:
          type: string
        now_utc:
          type: string
...
    PaginatedResponse:
      type: object
      properties:
        server_time_utc:
          type: string
          format: date-time
        events:
          type: array
          items:
            $ref: '#/components/schemas/Event'
        next_cursor:
          type: string
          nullable: true
...
    NewsBodyResponse:
      type: object
      properties:
        news_id:
          type: string
        title:
          type: string
        description:
          type: string
        article_url:
          type: string
        published_utc:
          type: string
        tickers:
          type: array
          items:
            type: string
        author:
          type: string
        keywords:
          type: array
          items:
            type: string
        image_url:
          type: string
        publisher:
          type: object
        insights:
          type: object
    NewsStatsResponse:
      type: object
      properties:
        total_count:
          type: integer
        date_range:
          type: object
          properties:
            earliest:
              type: string
              nullable: true
            latest:
              type: string
              nullable: true
        daily_counts:
          type: array
          items:
            type: object
            properties:
              date:
                type: string
              count:
                type: integer
        top_tickers:
          type: array
          items:
            type: object
            properties:
              ticker:
                type: string
              count:
                type: integer
        top_publishers:
          type: array
          items:
            type: object
            properties:
              publisher:
                type: string
              count:
                type: integer
        duplicate_stats:
          type: object
          properties:
            by_url:
              $ref: '#/components/schemas/DuplicateStats'
            by_title:
              $ref: '#/components/schemas/DuplicateStats'
    DuplicateStats:
      type: object
      properties:
        total:
          type: integer
        unique:
          type: integer
        duplicate_rate:
          type: number
          format: double
...